cloud9/was_service/requirements.txt
python-dotenv
flask
Pyjwt
boto3
pymysql